requirements.txt
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
joblib==1.4.2
kiwisolver==1.4.5
llvmlite==0.43.0
matplotlib==3.9.0
msgpack==1.0.8
numba==0.60.0
numpy==1.26.4
packaging==24.1
pillow==10.3.0
pynndescent==0.5.12
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
threadpoolctl==3.5.0
tqdm==4.66.4
umap-learn==0.5.6
hdbscan==0.8.37